Refactor and coordinate system overhaul
The map is now placed on the xy-plane, such that "height" is the z axis.
This way, first and second axes (x and y) are homogenous, where the
third axis is the odd one out. This means 2d coordinates are easier
converted to 3d, since the x and y directly translate between each
other.